Make `*Laplace` respect the model datatype.

+++ laplace/utils/utils.py
@@ -210,7 +210,8 @@ def symeig(M: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
     except RuntimeError:  # did not converge
         logging.info("SYMEIG: adding jitter, did not converge.")
         # use W L W^T + I = W (L + I) W^T
-        M = M + torch.eye(M.shape[0], device=M.device)
+        M = M + torch.eye(M.shape[0], device=M.device, dtype=M.dtype)
+
         try:
             L, W = torch.linalg.eigh(M, UPLO="U")
             L -= 1.0
@@ -219,6 +220,7 @@ def symeig(M: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
             stats = stats + f"min: {M.abs().min()}, mean: {M.abs().mean()}"
             logging.info(f"SYMEIG: adding jitter failed. Stats: {stats}")
             exit()
+
     # eigenvalues of symeig at least 0
     L = L.clamp(min=0.0)
     L = torch.nan_to_num(L)
